+++ example/android/app/src/main/kotlin/com/example/flutter_video_thumbnails_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.video_snapshot_generator_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}